+++ core/layer_pool_avg.h
@@ -99,12 +99,12 @@ class LayerPoolAvg: public LayerPool<Dtype> {
             uint32_t count = 0;
             for (uint32_t x = 0; x < Parent::filter_width_; ++x) {
               int32_t in_x = start_x + x;
-              if (in_x < 0 || in_x >= in_width) {
+              if (in_x < 0 || static_cast<uint32_t>(in_x) >= in_width) {
                 continue;
               }
               for (uint32_t y = 0; y < Parent::filter_height_; ++y) {
                 int32_t in_y = start_y + y;
-                if (in_y < 0 || in_y >= in_height) {
+                if (in_y < 0 || static_cast<uint32_t>(in_y) >= in_height) {
                   continue;
                 }
                 uint32_t in_index =
@@ -156,12 +156,12 @@ class LayerPoolAvg: public LayerPool<Dtype> {
               Parent::out_width_ + out_x;
             for (uint32_t x = 0; x < Parent::filter_width_; ++x) {
               int32_t in_x = start_x + x;
-              if (in_x < 0 || in_x >= in_width) {
+              if (in_x < 0 || static_cast<uint32_t>(in_x) >= in_width) {
                 continue;
               }
               for (uint32_t y = 0; y < Parent::filter_height_; ++y) {
                 int32_t in_y = start_y + y;
-                if (in_y < 0 || in_y >= in_height) {
+                if (in_y < 0 || static_cast<uint32_t>(in_y) >= in_height) {
                   continue;
                 }
                 uint32_t in_index =
